requirements.txt
torch>=1.7
torchvision>=0.8
tensorflow>=2.3
numpy
matplotlib
opencv-python
 adversarial-robustness-toolbox
pytorch-federated
timm
monai
tqdm